requirements.txt
pydantic==1.10.9
python-telegram-bot==20.6
SQLAlchemy==2.0.15
asyncpg==0.29.0
python-dotenv==1.0.0
aiohttp==3.9.0